requirements.txt
flask
werkzeug
torch
transformers
sentencepiece
Pillow
torchaudio
onnxruntime
numpy